services/python-services/search-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
azure-servicebus==7.11.4
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
prometheus-client==0.19.0
structlog==23.2.0

# Search & NLP
elasticsearch==8.11.0
opensearch-py==2.4.0
whoosh==2.7.4

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.2
tokenizers==0.15.0

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
joblib==1.3.2

# Text Processing & Analysis
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
textdistance==4.6.0
langdetect==1.0.9

# Search Analytics
rank-bm25==0.2.2
pyterrier==0.10.0

# Data Processing
python-multipart==0.0.6
httpx==0.25.2
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Monitoring and Health
prometheus-client==0.19.0
python-json-logger==2.0.7
opencensus-ext-azure==1.1.13

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Security & Encryption
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Caching & Performance
aiocache==0.12.2
async-lru==2.0.4